InvisibleInk-Experiments/InvisibleInk-Experiments-main/requirements.txt
gdown==5.2.0
numpy==2.2.4
pandas==2.2.3
scipy==1.15.2
scikit-learn==1.6.1
mauve-text==0.4.0
tqdm==4.67.1
torch==2.6.0
accelerate==1.5.2
transformers==4.57.0
datasets==3.4.1
sentence-transformers==3.4.1